Internal type safety
Expose Layout's getBaseline(). This is a Pango 1.22 function, so
update the requirements expressed by configure accordingly.

Change internal Pango scale constant to be a double rather than an int
so that we are less likely to make divide by integer errors in code
using it if we forget to cast the dividend to double.

Remove all the casts that are now unnecessary.

Fix the signature of FontDescription's getSize(). It needs to be a
double like everything else.

+++ src/bindings/org/gnome/pango/FontDescription.java
@@ -262,7 +262,7 @@ public final class FontDescription extends Boxed
      * 
      * @since 4.0.10
      */
-    public int getSize() {
+    public double getSize() {
         return PangoFontDescription.getSize(this) / Pango.SCALE;
     }
 }
